+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples.2/HorizontalFlow.jsx
@@ -115,7 +115,7 @@ const getNumberOfColumns = (activeBreakpoint) => {
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Icon',
+    displayName: 'Status',
     width: '60px',
     align: alignTypes.CENTER,
   },