+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Removed `sortIndicator` from the first column of Table With Sections example.
+  * Updated `terra-compact-interactive-list` examples to provide a meaningful name to the column with icons.
 
 * Added
   * Added examples and tests for `terra-table` to demonstrate auto layout.